Support Sqlite and MySQL in starting test server
