+++ server/constants.go
@@ -5,12 +5,13 @@ import "time"
 type GameMode string
 
 const (
-	Dev    GameMode = "dev"
-	DevSQL GameMode = "devSQL"
-	Prod   GameMode = "prod"
+	Dev       GameMode = "dev"
+	DevMySQL  GameMode = "devMySQL"
+	DevSQLite GameMode = "devSQLite"
+	Prod      GameMode = "prod"
 
-	SaveStateInterval       = time.Second * 10
-	DevSQLSaveStateInterval = time.Second
+	SaveStateInterval    = time.Second * 10
+	DevSaveStateInterval = time.Millisecond * 100
 
 	DefaultServerPort    = 9000
 	DefaultWebsocketPort = 9001
