+++ test/testutils/db.go
@@ -3,6 +3,9 @@ package testutils
 import (
 	"database/sql"
 	"fmt"
+	"github.com/stretchr/testify/assert"
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
 	"os"
 	"testing"
 
@@ -15,6 +18,7 @@ import (
 )
 
 var testSQLDSN string
+var testSQLiteDBPath = "test.db"
 
 func init() {
 	if err := godotenv.Load("../.env"); err != nil {
@@ -25,7 +29,7 @@ func init() {
 	txdb.Register("txdb", "mysql", testSQLDSN)
 }
 
-func SetupTestDB(t *testing.T, testGameID string, deleteTables bool, accessors map[interface{}]*state.TableBaseAccessor[any]) (*gamedb.MySQLSaveStateHandler, *gamedb.MySQLSaveTransactionHandler, *sql.DB) {
+func SetupMySQLTestDB(t *testing.T, testGameID string, deleteTables bool, accessors map[interface{}]*state.TableBaseAccessor[any]) (*gamedb.MySQLSaveStateHandler, *gamedb.MySQLSaveTransactionHandler, *sql.DB) {
 	var db *sql.DB
 	db, err := sql.Open("txdb", testSQLDSN)
 	if err != nil {
@@ -34,7 +38,7 @@ func SetupTestDB(t *testing.T, testGameID string, deleteTables bool, accessors m
 	require.Nil(t, db.Ping())
 
 	if deleteTables {
-		DeleteAllTables(t, db)
+		deleteAllTablesMySQL(t, db)
 	}
 
 	sqlDialector := mysql.New(mysql.Config{Conn: db})
@@ -44,7 +48,7 @@ func SetupTestDB(t *testing.T, testGameID string, deleteTables bool, accessors m
 	return mySQLSaveStateHandler, mySQLSaveTxHandler, db
 }
 
-func DeleteAllTables(t *testing.T, db *sql.DB) {
+func deleteAllTablesMySQL(t *testing.T, db *sql.DB) {
 	rows, err := db.Query("SHOW TABLES")
 	require.Nil(t, err)
 	defer rows.Close()
@@ -63,4 +67,77 @@ func DeleteAllTables(t *testing.T, db *sql.DB) {
 			fmt.Println("Failed to drop table", table, "err", err)
 		}
 	}
+}
+
+// setup local sqlite test db
+func SetupSQLiteTestDB(t *testing.T, testGameID string, deleteTables bool, accessors map[interface{}]*state.TableBaseAccessor[any]) (*gamedb.MySQLSaveStateHandler, *gamedb.MySQLSaveTransactionHandler, *sql.DB) {
+	db, err := sql.Open("sqlite3", testSQLiteDBPath)
+	if err != nil {
+		require.Nil(t, err)
+	}
+	require.Nil(t, db.Ping())
+
+	if deleteTables {
+		deleteAllTablesSQLite(t)
+	}
+
+	gormDB, err := gorm.Open(sqlite.Open(testSQLiteDBPath))
+
+	mySQLSaveStateHandler, mySQLSaveTxHandler, err := gamedb.SQLHandlersFromDialector(gormDB.Dialector, testGameID, accessors)
+	require.Nil(t, err)
+
+	return mySQLSaveStateHandler, mySQLSaveTxHandler, db
+}
+
+func ResetSQLiteTestDB() error {
+	dbFileName := testSQLiteDBPath
+
+	// Check if the file exists
+	if _, err := os.Stat(dbFileName); err == nil {
+		// File exists, so delete it
+		err := os.Remove(dbFileName)
+		if err != nil {
+			return err
+		}
+	}
+
+	// Create an empty file
+	file, err := os.Create(dbFileName)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+
+	return nil
+}
+
+// delete all tables in a sqlite db
+func deleteAllTablesSQLite(t *testing.T) {
+	db, err := gorm.Open(sqlite.Open(testSQLiteDBPath), &gorm.Config{})
+	if err != nil {
+		panic("failed to connect database")
+	}
+
+	// get list of table names
+	tableNames := getSQLiteTableNames(db)
+
+	// Iterate through the table names and drop each table
+	for _, tableName := range tableNames {
+		if err := db.Exec("DROP TABLE " + tableName + ";").Error; err != nil {
+			panic("Failed to drop table " + tableName + ": " + err.Error())
+		}
+	}
+
+	// verify that table names array is empty
+	updatedTableNames := getSQLiteTableNames(db)
+	assert.Equal(t, 0, len(updatedTableNames))
+
+}
+
+func getSQLiteTableNames(db *gorm.DB) []string {
+	var tableNames []string
+	if err := db.Raw("SELECT name FROM sqlite_master WHERE type='table';").Scan(&tableNames).Error; err != nil {
+		panic("Failed to fetch table names: " + err.Error())
+	}
+	return tableNames
 }
